fix leaked adb logic
